Ints to floats

+++ src/Game/GameEntity.cpp
@@ -9,18 +9,18 @@
 
 
 // Methods
-void GameEntity::horizontalMove(int move_x) {
+void GameEntity::horizontalMove(float move_x) {
   _position.x += move_x;
 }
 
 
 // Setters
-void GameEntity::setSize(int size_x, int size_y) {
+void GameEntity::setSize(float size_x, float size_y) {
   _size.x = size_x;
   _size.y = size_y;
 }
 
-void GameEntity::setPosition(int pos_x, int pos_y) {
+void GameEntity::setPosition(float pos_x, float pos_y) {
   _position.x = pos_x;
   _position.y = pos_y;
 }
